Split packer into individual packs
