+++ src/pyetm/models/packables/output_curves_pack.py
@@ -1,0 +1,28 @@
+import logging
+from typing import ClassVar, Any
+
+import pandas as pd
+
+from pyetm.models.packables.packable import Packable
+
+logger = logging.getLogger(__name__)
+
+
+class OutputCurvesPack(Packable):
+    key: ClassVar[str] = "output_curves"
+    sheet_name: ClassVar[str] = "OUTPUT_CURVES"
+
+    def _build_dataframe_for_scenario(self, scenario: Any, columns: str = "", **kwargs):  # type: ignore[override]
+        try:
+            series_list = list(scenario.all_output_curves())
+        except Exception as e:  # pragma: no cover
+            logger.warning(
+                "Failed extracting output curves for %s: %s", scenario.identifier(), e
+            )
+            return None
+        if not series_list:
+            return None
+        return pd.concat(series_list, axis=1)
+
+    def _to_dataframe(self, columns="", **kwargs) -> pd.DataFrame:  # type: ignore[override]
+        return self.build_pack_dataframe(columns=columns, **kwargs)
