+++ src/pyetm/models/scenario_packer.py
@@ -1,723 +1,20 @@
 import pandas as pd
 import logging
 from os import PathLike
-from pydantic import BaseModel, Field
-from typing import Optional, Dict, Any, Set, ClassVar
+from pydantic import BaseModel
+from typing import Optional, Dict, Any
 from xlsxwriter import Workbook
 
+from pyetm.models.packables.custom_curves_pack import CustomCurvesPack
+from pyetm.models.packables.inputs_pack import InputsPack
+from pyetm.models.packables.output_curves_pack import OutputCurvesPack
+from pyetm.models.packables.query_pack import QueryPack
+from pyetm.models.packables.sortable_pack import SortablePack
 from pyetm.models import Scenario
 from pyetm.models.custom_curves import CustomCurves
 from pyetm.utils.excel import add_frame_with_scenario_styling
 
 logger = logging.getLogger(__name__)
-
-
-class Packable(BaseModel):
-    # Use a proper default set and keep the type consistent
-    scenarios: Set["Scenario"] = Field(default_factory=set)
-    key: ClassVar[str] = "base_pack"
-    sheet_name: ClassVar[str] = "SHEET"
-
-    def add(self, *scenarios):
-        "Adds one or more scenarios to the packable"
-        self.scenarios.update(scenarios)
-
-    def discard(self, scenario):
-        "Removes a scenario from the pack"
-        self.scenarios.discard(scenario)
-
-    def clear(self):
-        # Reset to an empty set
-        self.scenarios.clear()
-
-    def summary(self) -> dict:
-        return {self.key: {"scenario_count": len(self.scenarios)}}
-
-    def to_dataframe(self, columns="") -> pd.DataFrame:
-        """Convert the pack into a dataframe"""
-        if len(self.scenarios) == 0:
-            return pd.DataFrame()
-
-        return self._to_dataframe(columns=columns)
-
-    def from_dataframe(self, df):
-        """Should parse the df and call correct setters on identified scenarios"""
-
-    def _to_dataframe(self, columns="", **kwargs) -> pd.DataFrame:
-        """Base implementation - kids should implement this"""
-        return pd.DataFrame()
-
-    def _find_by_identifier(self, identifier: str):
-        ident_str = str(identifier)
-        return next(
-            (s for s in self.scenarios if str(s.identifier()) == ident_str),
-            None,
-        )
-
-
-class InputsPack(Packable):
-    key: ClassVar[str] = "inputs"
-    sheet_name: ClassVar[str] = "PARAMETERS"
-
-    def __init__(self, **data):
-        super().__init__(**data)
-        self._scenario_short_names: Dict[str, str] = (
-            {}
-        )  # scenario_id -> short_name mapping
-
-    def set_scenario_short_names(self, scenario_short_names: Dict[str, str]):
-        """Set the mapping of scenario identifiers to their short names."""
-        self._scenario_short_names = scenario_short_names
-
-    def _find_by_short_name(self, short_name: str):
-        """Find scenario by its short name."""
-        short_name_str = str(short_name)
-        for scenario in self.scenarios:
-            scenario_id = str(scenario.identifier())
-            if self._scenario_short_names.get(str(scenario.id)) == short_name_str:
-                return scenario
-        return None
-
-    def _find_by_any(self, label: str):
-        """Attempt to resolve a label to a scenario by (in order):
-        1. short_name mapping
-        2. scenario.identifier() (title or id)
-        3. numeric id match
-        """
-        if label is None:
-            return None
-        label_str = str(label).strip()
-        # 1. short name
-        s = self._find_by_short_name(label_str)
-        if s is not None:
-            return s
-        # 2. identifier (title or id)
-        for scenario in self.scenarios:
-            if str(scenario.identifier()) == label_str:
-                return scenario
-        # 3. numeric id
-        try:
-            num = int(float(label_str))
-            for scenario in self.scenarios:
-                if scenario.id == num:
-                    return scenario
-        except Exception:
-            pass
-        return None
-
-    def _to_dataframe(self, columns="user", **kwargs):
-        return pd.concat(
-            [
-                scenario.inputs.to_dataframe(columns=columns)
-                for scenario in self.scenarios
-            ],
-            axis=1,
-            keys=[
-                self._scenario_short_names.get(
-                    str(scenario.identifier()), str(scenario.identifier())
-                )
-                for scenario in self.scenarios
-            ],
-        )
-
-    def _normalize_inputs_dataframe(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Normalize various inputs sheet shapes into canonical shape:
-        - Drop leading completely blank rows.
-        - Accept either:
-          (a) Two header rows (short_name row above a row containing 'user'), or
-          (b) Single header row of scenario labels (no explicit 'user' row) -> we fabricate 'user'.
-        - Support 1- or 2-level row index (input[, unit]).
-        Returns DataFrame with columns MultiIndex(label, 'user').
-        """
-        df = df.dropna(how="all")
-        if df.empty:
-            return df
-
-        # Locate the row containing 'user' (case-insensitive)
-        user_row_pos = None
-        for pos, (_, row) in enumerate(df.iterrows()):
-            if any(isinstance(v, str) and v.strip().lower() == "user" for v in row):
-                user_row_pos = pos
-                break
-
-        single_header = user_row_pos is None
-        if single_header:
-            header_start = 0
-            header_end = 0
-        else:
-            header_start = max(user_row_pos - 1, 0)
-            header_end = user_row_pos
-
-        headers = df.iloc[header_start : header_end + 1].astype(str)
-        data = df.iloc[header_end + 1 :].copy()
-
-        if single_header:
-            # Build columns from single header row; fabricate second level 'user'
-            col_level0 = headers.iloc[0].values
-            # First one (index col) often blank -> keep for alignment
-            data.columns = col_level0
-            # Identify index columns: left-most non-scenario columns (we take first (and optional second) column as index parts)
-            index_candidates = list(data.columns[:2])  # heuristic
-            # If second column values are all numeric -> treat as scenario column, so only first column is index
-            second_is_numeric = False
-            if len(index_candidates) > 1:
-                sample = data[index_candidates[1]].dropna().head(5)
-                if not sample.empty and all(
-                    pd.to_numeric(sample, errors="coerce").notna()
-                ):
-                    second_is_numeric = True
-            if second_is_numeric:
-                idx_cols = [index_candidates[0]]
-            else:
-                idx_cols = index_candidates
-            input_col = idx_cols[0]
-            unit_col = idx_cols[1] if len(idx_cols) > 1 else None
-            scenario_cols = [
-                c for c in data.columns if c not in idx_cols and str(c).strip() != ""
-            ]
-            # Build index
-            input_series = data[input_col].astype(str)
-            if unit_col is not None:
-                unit_series = data[unit_col].astype(str)
-                index = pd.MultiIndex.from_arrays(
-                    [input_series.values, unit_series.values], names=["input", "unit"]
-                )
-            else:
-                index = pd.Index(input_series.values, name="input")
-            canonical = data[scenario_cols].copy()
-            canonical.index = index
-            # Fabricate second level
-            canonical.columns = pd.MultiIndex.from_arrays(
-                [canonical.columns, ["user"] * len(canonical.columns)]
-            )
-            return canonical
-
-        # Two-row header path (original logic)
-        data.columns = pd.MultiIndex.from_arrays(
-            [headers.iloc[0].values, headers.iloc[1].values]
-        )
-
-        idx_cols = [
-            col
-            for col in data.columns
-            if not (isinstance(col[1], str) and col[1].strip().lower() == "user")
-        ]
-        if len(idx_cols) == 0:
-            input_col = data.columns[0]
-            unit_col = None
-        else:
-            input_col = idx_cols[0]
-            unit_col = idx_cols[1] if len(idx_cols) > 1 else None
-
-        input_series = data[input_col].astype(str)
-        if unit_col is not None:
-            unit_series = data[unit_col].astype(str)
-            index = pd.MultiIndex.from_arrays(
-                [input_series.values, unit_series.values], names=["input", "unit"]
-            )
-        else:
-            index = pd.Index(input_series.values, name="input")
-
-        keep_cols = [
-            c
-            for c in data.columns
-            if c not in {input_col} and (unit_col is None or c != unit_col)
-        ]
-        canonical = data[keep_cols]
-        canonical.index = index
-
-        # Ensure second level equals 'user'
-        if isinstance(canonical.columns, pd.MultiIndex):
-            lvl1 = canonical.columns.get_level_values(1)
-            if not all(
-                isinstance(v, str) and v.strip().lower() == "user" for v in lvl1
-            ):
-                canonical.columns = pd.MultiIndex.from_arrays(
-                    [
-                        canonical.columns.get_level_values(0),
-                        ["user"] * len(canonical.columns),
-                    ]
-                )
-        else:
-            canonical.columns = pd.MultiIndex.from_arrays(
-                [
-                    canonical.columns,
-                    ["user"] * len(canonical.columns),
-                ]
-            )
-        return canonical
-
-    def from_dataframe(self, df):
-        """
-        Sets the inputs on the scenarios from the packed df (comes from excel)
-        Tolerates optional unit column and leading blank rows.
-        Uses short_name for scenario identification; falls back to identifier/title or id.
-        """
-        if df is None or getattr(df, "empty", False):
-            return
-
-        try:
-            df = self._normalize_inputs_dataframe(df)
-        except Exception as e:
-            logger.warning("Failed to normalize inputs sheet: %s", e)
-            return
-
-        if df is None or df.empty:
-            return
-
-        # Now df has columns MultiIndex (label, 'user') where label may be short_name, title, or id
-        labels = df.columns.get_level_values(0).unique()
-
-        for label in labels:
-            scenario = self._find_by_any(label)
-            if scenario is None:
-                logger.warning(
-                    "Could not find scenario for parameters column label '%s' (not a short_name/title/id)",
-                    label,
-                )
-                continue
-
-            scenario_df = df[label]
-            # Ensure DataFrame with a 'user' column
-            if isinstance(scenario_df, pd.Series):
-                scenario_df = scenario_df.to_frame(name="user")
-            else:
-                if list(scenario_df.columns) != ["user"]:
-                    scenario_df = scenario_df.copy()
-                    first_col = scenario_df.columns[0]
-                    scenario_df = scenario_df.rename(columns={first_col: "user"})
-
-            try:
-                scenario.set_user_values_from_dataframe(scenario_df)
-            except Exception as e:
-                logger.warning(
-                    "Failed setting inputs for scenario '%s' from column label '%s': %s",
-                    scenario.identifier(),
-                    label,
-                    e,
-                )
-
-
-class QueryPack(Packable):
-    key: ClassVar[str] = "gquery"
-    sheet_name: ClassVar[str] = "GQUERIES_RESULTS"
-
-    def _to_dataframe(
-        self, columns="future", **kwargs
-    ) -> pd.DataFrame:  # Make sure **kwargs is here
-        if not self.scenarios:
-            return pd.DataFrame()
-
-        return pd.concat(
-            [scenario.results(columns=columns) for scenario in self.scenarios],
-            axis=1,
-            keys=[scenario.identifier() for scenario in self.scenarios],
-            copy=False,
-        )
-
-    def _normalize_queries_dataframe(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Normalize a GQUERIES sheet into a simple shape:
-        - Drop leading completely blank rows and columns.
-        - Detect 1 or 2 header rows.
-        - Return a DataFrame with columns = scenario identifiers and rows listing gquery keys.
-        We ignore any leftmost index/helper columns whose header is empty or looks like a label.
-        """
-        if df is None:
-            return pd.DataFrame()
-
-        # Drop completely empty rows/cols
-        df = df.dropna(how="all")
-        if df.empty:
-            return df
-        df = df.dropna(axis=1, how="all")
-
-        # Find non-empty rows for potential headers
-        non_empty_rows = [
-            i for i, (_, r) in enumerate(df.iterrows()) if not r.isna().all()
-        ]
-        if not non_empty_rows:
-            return pd.DataFrame()
-
-        # Heuristic: if the second non-empty row contains the word 'gquery'/'gqueries', use 2 header rows
-        header_rows = 1
-        if len(non_empty_rows) > 1:
-            second = (
-                df.iloc[non_empty_rows[1]].astype(str).str.strip().str.lower().tolist()
-            )
-            if any(val in {"gquery", "gqueries", "key", "queries"} for val in second):
-                header_rows = 2
-
-        header_start = non_empty_rows[0]
-        header_end = header_start + header_rows - 1
-        headers = df.iloc[header_start : header_end + 1].astype(str)
-        data = df.iloc[header_end + 1 :].copy()
-
-        # Assign columns: MultiIndex if 2 header rows else single level
-        if header_rows == 2:
-            cols = pd.MultiIndex.from_arrays(
-                [headers.iloc[0].values, headers.iloc[1].values]
-            )
-        else:
-            cols = pd.Index(headers.iloc[0].values)
-        data.columns = cols
-
-        def _is_empty(v):
-            return (
-                (not isinstance(v, str))
-                or (v.strip() == "")
-                or (v.strip().lower() == "nan")
-            )
-
-        def _is_helper_label(v):
-            return isinstance(v, str) and v.strip().lower() in {
-                "gquery",
-                "gqueries",
-                "queries",
-                "key",
-            }
-
-        # Build a simple DataFrame with columns = scenario identifiers
-        if isinstance(data.columns, pd.MultiIndex):
-            keep = [
-                c
-                for c in data.columns
-                if not _is_empty(c[0]) and not _is_helper_label(c[0])
-            ]
-            data = data[keep]
-            # Collapse to single level (identifier)
-            data.columns = [c[0] for c in data.columns]
-        else:
-            keep = [
-                c
-                for c in data.columns
-                if isinstance(c, str) and not _is_empty(c) and not _is_helper_label(c)
-            ]
-            data = data[keep]
-
-        # Ensure string values and drop rows that are completely blank across all kept columns
-        for c in data.columns:
-            data[c] = data[c].apply(lambda x: None if pd.isna(x) else (str(x).strip()))
-        data = data.dropna(how="all")
-
-        return data
-
-    def from_dataframe(self, df: pd.DataFrame):
-        """Collect gquery keys for each scenario from a GQUERIES sheet and attach them.
-        The sheet is expected to have one column per scenario (by identifier/title),
-        with each row containing a gquery key. Blank rows are ignored.
-        """
-        if df is None or getattr(df, "empty", False):
-            return
-
-        try:
-            df = self._normalize_queries_dataframe(df)
-        except Exception as e:
-            logger.warning("Failed to normalize gqueries sheet: %s", e)
-            return
-
-        if df is None or df.empty:
-            return
-
-        for identifier in df.columns:
-            scenario = self._find_by_identifier(identifier)
-            if scenario is None:
-                logger.warning(
-                    "Could not find scenario for identifier '%s'", identifier
-                )
-                continue
-
-            # Extract non-empty keys, preserve order, remove duplicates while preserving order
-            values = [
-                v
-                for v in df[identifier].tolist()
-                if isinstance(v, str) and v.strip() != ""
-            ]
-            seen = set()
-            keys = []
-            for v in values:
-                if v not in seen:
-                    seen.add(v)
-                    keys.append(v)
-
-            if keys:
-                try:
-                    scenario.add_queries(keys)
-                except Exception as e:
-                    logger.warning("Failed to add gqueries to '%s': %s", identifier, e)
-
-
-class SortablePack(Packable):
-    key: ClassVar[str] = "sortables"
-    sheet_name: ClassVar[str] = "SORTABLES"
-
-    def _to_dataframe(self, columns="", **kwargs) -> pd.DataFrame:
-        """Pack sortables data for all scenarios with multi-index support"""
-        if not self.scenarios:
-            return pd.DataFrame()
-
-        sortables_dfs = []
-        scenario_keys = []
-
-        for scenario in self.scenarios:
-            df = scenario.sortables.to_dataframe()
-            if not df.empty:
-                sortables_dfs.append(df)
-                scenario_keys.append(scenario.identifier())
-
-        if not sortables_dfs:
-            return pd.DataFrame()
-
-        return pd.concat(
-            sortables_dfs,
-            axis=1,
-            keys=scenario_keys,
-        )
-
-    def _normalize_sortables_dataframe(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Normalize various sortables sheet shapes.
-        Assumptions:
-        - Two header rows: first row = scenario identifier/title, second row = sortable name.
-        - Leading blank rows may exist and are ignored.
-        - Optional leftmost index column(s) are present with empty first header cell(s). They are dropped.
-        Returns a DataFrame with columns MultiIndex(identifier, sortable_name) and
-        simple Index rows with order positions.
-        """
-        # Drop completely empty rows
-        df = df.dropna(how="all")
-        if df.empty:
-            return df
-
-        # Find the first two non-empty rows -> headers
-        non_empty_idx = [
-            i for i, (_, r) in enumerate(df.iterrows()) if not r.isna().all()
-        ]
-        if not non_empty_idx:
-            return pd.DataFrame()
-        header0_pos = non_empty_idx[0]
-        header1_pos = non_empty_idx[1] if len(non_empty_idx) > 1 else header0_pos + 1
-
-        headers = df.iloc[[header0_pos, header1_pos]].astype(str)
-        data = df.iloc[header1_pos + 1 :].copy()
-
-        # Build MultiIndex columns
-        col_level0 = headers.iloc[0].values
-        col_level1 = headers.iloc[1].values
-        columns = pd.MultiIndex.from_arrays([col_level0, col_level1])
-        data.columns = columns
-
-        # Drop columns where identifier (level 0) is missing/empty, and drop any
-        # column which is clearly the helper/index label (e.g. level1 == 'sortables').
-        def _is_empty(v):
-            return (
-                (not isinstance(v, str))
-                or (v.strip() == "")
-                or (v.strip().lower() == "nan")
-            )
-
-        def _is_helper_label(v):
-            return isinstance(v, str) and v.strip().lower() in {"sortables"}
-
-        keep_cols = [
-            c
-            for c in data.columns
-            if not _is_empty(c[0]) and not _is_helper_label(c[0])
-        ]
-        data = data[keep_cols]
-        # Collapse to single level (identifier)
-        data.columns = [c[0] for c in data.columns]
-
-        # Ensure string values and drop rows that are completely blank across all kept columns
-        for c in data.columns:
-            data[c] = data[c].apply(lambda x: None if pd.isna(x) else (str(x).strip()))
-        data = data.dropna(how="all")
-
-        return data
-
-    def from_dataframe(self, df: pd.DataFrame):
-        """Unpack and update sortables for each scenario from the sheet.
-        The sheet may contain optional leading blank rows and optional index columns.
-        """
-        if df is None or getattr(df, "empty", False):
-            return
-
-        try:
-            df = self._normalize_sortables_dataframe(df)
-        except Exception as e:
-            logger.warning("Failed to normalize sortables sheet: %s", e)
-            return
-
-        if df is None or df.empty or not isinstance(df.columns, pd.MultiIndex):
-            return
-
-        identifiers = df.columns.get_level_values(0).unique()
-        for identifier in identifiers:
-            scenario = self._find_by_identifier(identifier)
-            if scenario is None:
-                logger.warning(
-                    "Could not find scenario for identifier '%s'", identifier
-                )
-                continue
-
-            block = df[identifier]
-            try:
-                scenario.set_sortables_from_dataframe(block)
-            except Exception as e:
-                logger.warning("Failed to update sortables for '%s': %s", identifier, e)
-
-
-class CustomCurvesPack(Packable):
-    key: ClassVar[str] = "custom_curves"
-    sheet_name: ClassVar[str] = "CUSTOM_CURVES"
-
-    def _to_dataframe(self, columns="", **kwargs) -> pd.DataFrame:
-        """Pack custom curves data for all scenarios with multi-index support"""
-        if not self.scenarios:
-            return pd.DataFrame()
-
-        curves_dfs = []
-        scenario_keys = []
-
-        for scenario in self.scenarios:
-            series_list = list(scenario.custom_curves_series())
-            if len(series_list) > 0:
-                df = pd.concat(series_list, axis=1)
-                curves_dfs.append(df)
-                scenario_keys.append(scenario.identifier())
-
-        if not curves_dfs:
-            return pd.DataFrame()
-
-        return pd.concat(
-            curves_dfs,
-            axis=1,
-            keys=scenario_keys,
-        )
-
-    def _normalize_curves_dataframe(self, df: pd.DataFrame) -> pd.DataFrame:
-        """Normalize custom curves sheet shapes.
-        Assumptions:
-        - Two header rows: first row = scenario identifier/title, second row = curve key.
-        - Leading blank rows may exist and are ignored.
-        - Optional leftmost index column(s) are present with empty first header cell(s); they are dropped.
-        - Rows are the hourly values (0..8759) or arbitrary length; we keep as-is.
-        Returns a DataFrame with columns MultiIndex(identifier, curve_key) and numeric rows.
-        """
-        df = df.dropna(how="all")
-        if df.empty:
-            return df
-
-        # Find header rows
-        non_empty_idx = [
-            i for i, (_, r) in enumerate(df.iterrows()) if not r.isna().all()
-        ]
-        if not non_empty_idx:
-            return pd.DataFrame()
-        header0_pos = non_empty_idx[0]
-        header1_pos = non_empty_idx[1] if len(non_empty_idx) > 1 else header0_pos + 1
-
-        headers = df.iloc[[header0_pos, header1_pos]].astype(str)
-        data = df.iloc[header1_pos + 1 :].copy()
-
-        # Assign columns
-        columns = pd.MultiIndex.from_arrays(
-            [headers.iloc[0].values, headers.iloc[1].values]
-        )
-        data.columns = columns
-
-        # Drop non-scenario columns where identifier is empty and drop helper/index
-        # columns where level1 looks like a label (e.g., 'sortables')
-        def _is_empty(v):
-            return (
-                (not isinstance(v, str))
-                or (v.strip() == "")
-                or (v.strip().lower() == "nan")
-            )
-
-        def _is_helper_label(v):
-            return isinstance(v, str) and v.strip().lower() in {"sortables"}
-
-        keep_cols = [
-            c
-            for c in data.columns
-            if not _is_empty(c[0]) and not _is_helper_label(c[1])
-        ]
-        canonical = data[keep_cols].copy()
-
-        # Reset index to numeric starting at 0
-        canonical.reset_index(drop=True, inplace=True)
-        return canonical
-
-    def from_dataframe(self, df: pd.DataFrame):
-        """Unpack and update custom curves for each scenario.
-        Sheet may contain leading blank rows and optional index columns.
-        """
-        if df is None or getattr(df, "empty", False):
-            return
-
-        try:
-            df = self._normalize_curves_dataframe(df)
-        except Exception as e:
-            logger.warning("Failed to normalize custom curves sheet: %s", e)
-            return
-
-        if df is None or df.empty or not isinstance(df.columns, pd.MultiIndex):
-            return
-
-        identifiers = df.columns.get_level_values(0).unique()
-        for identifier in identifiers:
-            scenario = self._find_by_identifier(identifier)
-            if scenario is None:
-                logger.warning(
-                    "Could not find scenario for identifier '%s'", identifier
-                )
-                continue
-
-            block = df[identifier]
-            # Build a CustomCurves collection from the block columns
-            try:
-                curves = CustomCurves._from_dataframe(block, scenario_id=scenario.id)
-            except Exception as e:
-                logger.warning(
-                    "Failed to build custom curves for '%s': %s", identifier, e
-                )
-                continue
-
-            # Validate and upload
-            try:
-                scenario.update_custom_curves(curves)
-            except Exception as e:
-                logger.warning(
-                    "Failed to update custom curves for '%s': %s", identifier, e
-                )
-
-
-class OutputCurvesPack(Packable):
-    key: ClassVar[str] = "output_curves"
-    sheet_name: ClassVar[str] = "OUTPUT_CURVES"
-
-    def _to_dataframe(self, columns="", **kwargs) -> pd.DataFrame:
-        """Pack output curves data for all scenarios with multi-index support"""
-        if not self.scenarios:
-            return pd.DataFrame()
-
-        curves_dfs = []
-        scenario_keys = []
-
-        for scenario in self.scenarios:
-            series_list = list(scenario.all_output_curves())
-            if len(series_list) > 0:
-                df = pd.concat(series_list, axis=1)
-                curves_dfs.append(df)
-                scenario_keys.append(scenario.identifier())
-
-        if not curves_dfs:
-            return pd.DataFrame()
-
-        return pd.concat(
-            curves_dfs,
-            axis=1,
-            keys=scenario_keys,
-        )
 
 
 class ScenarioPacker(BaseModel):
